app/pyproject.toml
[tool.poetry]
name = "take-home"
version = "0.1.0"
description = "Basic Django web application for take home assignment."
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.3"
psycopg2-binary = "^2.9.6"
djangorestframework = "^3.14.0"
pytest = "^8.0.0"
pytest-django = "^4.7.0"
black = "^24.1.1"
pyyaml = "^6.0.1"
django-rest-swagger = "^2.2.0"
uritemplate = "^4.1.1"
djangorestframework-simplejwt = "^5.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
pytest = "pytest"